# controller/admin_login.go
package controller

import (
	"encoding/json"
	"github.com/e421083458/golang_common/lib"
	"github.com/gin-gonic/contrib/sessions"
	"github.com/gin-gonic/gin"
	"go_gateway/dao"
	"go_gateway/dto"
	"go_gateway/middleware"
	"go_gateway/public"
	"time"
)

type AdminLoginController struct {

}

func AdminLoginRegister(group *gin.RouterGroup)  {
	adminLogin :=&AdminLoginController{}
	group.POST("/login",adminLogin.AdminLogin)
	group.GET("/logout",adminLogin.AdminLoginOut)
}

// AdminLogin godoc
// @Summary 管理员登录
// @Description 管理员登录
// @Tags 管理员接口
// @ID /admin_login/login
// @Accept  json
// @Produce  json
// @Param body body dto.AdminLoginInput true "body"
// @Success 200 {object} middleware.Response{data=dto.AdminLoginOutput} "success"
// @Router /admin_login/login [post]
func (adminlogin *AdminLoginController)AdminLogin(c *gin.Context)  {
	params := &dto.AdminLoginInput{}
	if err:= params.BindValidParam(c);err!=nil{
		middleware.ResponseError(c,2000,err)
		return
	}
	tx,err:=lib.GetGormPool("default")
	if err != nil {
		middleware.ResponseError(c,2001,err)
		return
	}
	admin := &dao.Admin{}
	admin,err= admin.LoginCheck(c,tx,params)
	if err != nil {
		middleware.ResponseError(c,2002,err)
		return
	}
	sessionJSON, err := json.Marshal(&dto.AdminSessionInfo{
		ID:        admin.ID,
		UserName:  admin.UserName,
		LoginTime: time.Now(),
	})
	if err != nil {
		middleware.ResponseError(c,2003,err)
		return
	}

	session:=sessions.Default(c)
	session.Set(public.AdminSessionInfoKey,string(sessionJSON))
	session.Save()


	out:=&dto.AdminLoginOutput{Token: admin.UserName}
	middleware.ResponseSuccess(c,out)
}





// AdminLoginOut godoc
// @Summary 管理员退出
// @Description 管理员退出
// @Tags 管理员接口
// @ID /admin_login/logout
// @Accept  json
// @Produce  json
// @Success 200 {object} middleware.Response{data=string} "success"
// @Router /admin_login/logout [get]
func (adminlogin *AdminLoginController)AdminLoginOut(c *gin.Context)  {


	session:=sessions.Default(c)
	session.Delete(public.AdminSessionInfoKey)
	session.Save()



	middleware.ResponseSuccess(c,"")
}